fix: LottoResult.totalPrizeMoney 타입을 Long으로 변경
Int.MAX_VALUE 를 넘어서 overflow가 되면 수익률이 음수로 되는 버그를 수정

+++ src/main/kotlin/lotto/domain/LottoResult.kt
@@ -1,10 +1,10 @@
 package lotto.domain
 
 class LottoResult private constructor(private val result: Map<Prize, Int>) : Map<Prize, Int> by result {
-    private val totalPrizeMoney: Int
+    private val totalPrizeMoney: Long
         get() {
             return result.map {
-                it.key.money * it.value
+                it.key.money.toLong() * it.value
             }.sum()
         }
